requirements.txt
pandas
numpy
azure-storage-blob
pyyaml
scikit-learn
python-box[all]~=7.0
ensure
# -e .